+++ @geogebra/solver-sdk/test/parsers.test.ts
@@ -1465,7 +1465,7 @@ describe('Solver Parser Unit Tests', () => {
     ]);
   });
 
-  describe('Trigonometric functions', () => {
+  describe('Trigonometric functions simple operations', () => {
     testCases([
       {
         solver: 'sin x',
@@ -1506,6 +1506,8 @@ describe('Solver Parser Unit Tests', () => {
         latex: [
           '\\sin{\\left(x\\right)}',
           '{{\\mathrm{sin}}\\left(x\\right)}', // ggb keyboard output
+          '{\\mathrm{sin}\\left(x\\right)}',
+          '\\mathrm{sin}\\left(x\\right)', // ggb keyboard output
         ],
       },
       {
@@ -1632,10 +1634,6 @@ describe('Solver Parser Unit Tests', () => {
         ],
       },
       {
-        // TODO: add "json" and "solver" syntax for below once we agree on their syntax's
-        latex: ['\\sin^{2} x', '\\sin^2 x', '{{\\mathrm{sin}}^{2} x}'],
-      },
-      {
         solver: 'tan (x)',
         json: {
           type: 'Tan',
@@ -1651,6 +1649,123 @@ describe('Solver Parser Unit Tests', () => {
           '\\tan{\\left(x\\right)}',
           '{{\\mathrm{tan}}\\left(x\\right)}', // ggb keyboard output
         ],
+      },
+    ]);
+  });
+
+  describe('Power of trigonometric functions', () => {
+    testCases([
+      {
+        solver: '[sin ^ 2] (x)',
+        json: {
+          type: 'Power',
+          operands: [
+            {
+              type: 'Sin',
+              operands: [
+                {
+                  type: 'Variable',
+                  value: 'x',
+                  decorators: ['RoundBracket'],
+                },
+              ],
+              powerInside: true,
+            },
+            integer('2'),
+          ],
+        },
+        latex: [
+          '\\sin^{2}{\\left(x\\right)}',
+          '\\sin^2 {(x)}',
+          // below is "expected" ggb keyboard output, current keyboard output is '{{\mathrm{sin^{2}}}\left(x\right)}'
+          '{{\\mathrm{sin}}^{2} \\left(x\\right)}',
+          '{\\mathrm{sin}^{2} \\left(x\\right)}',
+          '\\mathrm{sin}^{2} \\left(x\\right)',
+        ],
+      },
+      {
+        solver: '[arctan ^ 2] (x)',
+        json: {
+          type: 'Power',
+          operands: [
+            {
+              type: 'Arctan',
+              operands: [
+                {
+                  type: 'Variable',
+                  value: 'x',
+                  decorators: ['RoundBracket'],
+                },
+              ],
+              powerInside: true,
+            },
+            integer('2'),
+          ],
+        },
+        latex: ['\\arctan^{2}{\\left(x\\right)}', '\\arctan^2 {(x)}'],
+      },
+      {
+        solver: '[cosh ^ 3] (x)',
+        json: {
+          type: 'Power',
+          operands: [
+            {
+              type: 'Cosh',
+              operands: [
+                {
+                  type: 'Variable',
+                  value: 'x',
+                  decorators: ['RoundBracket'],
+                },
+              ],
+              powerInside: true,
+            },
+            integer('3'),
+          ],
+        },
+        latex: ['\\cosh^{3}{\\left(x\\right)}', '\\cosh^3 {(x)}'],
+      },
+    ]);
+  });
+
+  describe('Inverse trigonometric functions', () => {
+    testCases([
+      {
+        solver: '[sin ^ -1] (x)',
+        json: {
+          type: 'Arcsin',
+          operands: [
+            {
+              type: 'Variable',
+              value: 'x',
+              decorators: ['RoundBracket'],
+            },
+          ],
+          inverseNotation: 'superscript',
+        },
+        latex: ['\\sin^{-1}{\\left(x\\right)}', '{{\\mathrm{sin}}^{-1}\\left(x\\right)}'],
+      },
+      {
+        solver: '[sinh ^ -1] (x)',
+        json: {
+          type: 'Arsinh',
+          operands: [
+            {
+              type: 'Variable',
+              value: 'x',
+              decorators: ['RoundBracket'],
+            },
+          ],
+          inverseNotation: 'superscript',
+        },
+        latex: [
+          '\\sinh^{-1}{\\left(x\\right)}',
+          '{{\\mathrm{sinh}}^{-1}\\left(x\\right)}',
+        ],
+      },
+      {
+        // inverse of an inverse function is treated and parsed the same way as the function currently
+        latex: ['\\arcsin^{-1}{x}', '\\sin x'],
       },
     ]);
   });