+++ engine/src/test/kotlin/engine/expressions/TestToJson.kt
@@ -24,6 +24,40 @@ class TestToJson {
     )
 
     @Test
+    fun sineCubedWithCubeInsideTest() = test(
+        "[sin ^ 3] (x)",
+        """{
+            "type": "Power",
+            "operands": [
+                {
+                    "type": "Sin",
+                    "operands": [
+                        {"type": "Variable", "value": "x", "decorators": ["RoundBracket"]}
+                    ],
+                    "inverseNotation": "arcPrefix",
+                    "powerNotation": "inside"
+                },
+                {"type": "Integer", "value": "3"}
+            ]
+        }
+        """,
+    )
+
+    @Test
+    fun sinhInverseTest() = test(
+        "[sinh ^ -1] x",
+        """{
+            "type": "Arsinh",
+            "operands": [
+                {"type": "Variable", "value": "x"}
+            ],
+            "inverseNotation": "superscript",
+            "powerNotation": "outside"
+        }
+        """,
+    )
+
+    @Test
     fun productTest() = test(
         "2x*3*y",
         """{